feat(comments-sentences): create a rule to enforce comments sentences
A comment must start with a space, then an uppercase letter, a number, a
quote (single or double) or an "_" (or any ignored pattern among
"eslint", "istanbul","jscs", "jshint", "globals?", "exported", "@.+")
and end with a punctuation (".", "!" or "?").
The comment can spans over multiple line (using multi-line comments
"/** */" or "/* */", "//" are just one line comments).
URL are not checked against this rule either.

It's possible to set `ignoreInlineComments` to... ignore inline comments
(you don't say!!!).
It's possible to add more `ignorePattern` to exclude more comments
format from this rule.